rectangle rendered to screen, engine works for relative locations of objects

+++ src/gui/engine/mesh.h
@@ -4,7 +4,7 @@
 #include "defines.h"
 
 typedef struct {
-	f16 x, y, width, height;
+	f32 x, y, width, height;
 } M_Rect; // rectangular bounds
 
 typedef struct {
