+++ src/gui/engine/render.h
@@ -15,7 +15,7 @@ void render_init(void);
 void render_uninit(Mesh* meshes, const u32 count);
 
 // mesh values updated with VAO and VBO
-void render_load_mesh(const f16* vertices, const u32 triangle_count, Mesh* mesh);
+void render_load_mesh(const f32* vertices, const u32 triangle_count, Mesh* mesh);
 void render(Mesh* meshes, u32 mesh_count);
 
 #endif
